requirements.txt
streamlit==1.28.0
pandas==2.1.1
sqlalchemy==2.0.23
cryptography==41.0.7
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
pdfkit==1.0.0
fpdf2==2.7.5
python-dotenv==1.0.0
pydantic==2.5.0
stripe==7.6.0
requests==2.31.0
